backend_old/requirements.txt
fastapi==0.111.0
pydantic==2.7.1
pydantic_settings==2.2.1
SQLAlchemy==2.0.30
uvicorn==0.29.0
asyncpg==0.29.0
fastapi-limiter==0.1.6
redis==5.0.7
passlib==1.7.4
python-jose==3.3.0
pytz